src/gpt/requirements.txt
flask
gunicorn
llama-cpp-python==0.2.70